Configuration for blue/green deployments w/ Docker
This repository contains a `docker-compose.yml` configuration that
builds and starts an HAProxy container load-balancing between two
containers running an application (called `sample_service` and written
in Go on that precise example).

The interesting part of it is in the `./deploy_blue_green.sh` script. It
basically performs a blue/green deployment, leveraging HAProxy's health
check to make sure that the new Docker image is functional.

Perspectives:
  * Externalize parts that depend on `sample_service` in environment
    variables so that this script can be used to deploy almost anything
    we'd like to.
  * Optimize HAProxy's config.
  * Investigate what happens for pending requests already allocated to a
    container (say blue) if the latter sutdowns before finishing to
    respond. (There's two options: it re-sends the request to green or
    it forwards the answer to the client. We'd like to have the first
    behaviour ofc.)