Add timeout when HAProxy health check fails
Now when the HAproxy health checks for more than 30seconds after the new blue
image launch, the deployment script exits with error code 12.

The way, we're guaranteed that deployments don't take more than 30 seconds and
even better, we should be warned when the health check fails so that the commit
responsible for breaking the bootstrap process can be reverted or fixed.